feat(persistence): add PostgreSQL persistence library with EF Core integration
Implement comprehensive PostgreSQL persistence solution including:
- Base context with audit fields and soft delete support
- Entity configurations for different entity types
- ADO.NET utilities for parameter handling and SQL sanitization
- Timezone and string extension utilities
- Exception handling and translation
- Unit of Work pattern implementation
- Data reader and table mapping utilities
- Dependency injection setup
- Project configuration and README
