+++ Acontplus.DotNet.Libs.sln
@@ -44,9 +44,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Acontplus.TestInfrastructur
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Acontplus.TestApi", "apps\src\Acontplus.TestApi\Acontplus.TestApi.csproj", "{18821081-AC99-58B3-FD8A-08E0C0F9F22D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Acontplus.Persistence.Postgres", "src\Acontplus.Persistence.Postgres\Acontplus.Persistence.Postgres.csproj", "{D6B82683-7DAE-4804-AE1F-477B892E197D}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Acontplus.Persistence.Common", "src\Acontplus.Persistence.Common\Acontplus.Persistence.Common.csproj", "{97E0521A-3EDD-781D-CE28-22CC465D76FA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Acontplus.Persistence.PostgreSQL", "src\Acontplus.Persistence.PostgreSQL\Acontplus.Persistence.PostgreSQL.csproj", "{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -238,18 +238,6 @@ Global
 		{18821081-AC99-58B3-FD8A-08E0C0F9F22D}.Release|x64.Build.0 = Release|Any CPU
 		{18821081-AC99-58B3-FD8A-08E0C0F9F22D}.Release|x86.ActiveCfg = Release|Any CPU
 		{18821081-AC99-58B3-FD8A-08E0C0F9F22D}.Release|x86.Build.0 = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|x64.Build.0 = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Debug|x86.Build.0 = Debug|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|x64.ActiveCfg = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|x64.Build.0 = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|x86.ActiveCfg = Release|Any CPU
-		{D6B82683-7DAE-4804-AE1F-477B892E197D}.Release|x86.Build.0 = Release|Any CPU
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -262,6 +250,18 @@ Global
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Release|x64.Build.0 = Release|Any CPU
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Release|x86.ActiveCfg = Release|Any CPU
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA}.Release|x86.Build.0 = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|x64.Build.0 = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Debug|x86.Build.0 = Debug|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|x64.ActiveCfg = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|x64.Build.0 = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|x86.ActiveCfg = Release|Any CPU
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -282,8 +282,8 @@ Global
 		{DB059BD9-CF37-527A-EB68-A6527ED40B88} = {802C95C5-F8AA-4136-A298-261105286C93}
 		{D8D945A8-EDC7-42D4-9353-7A7EE8DE4F58} = {802C95C5-F8AA-4136-A298-261105286C93}
 		{18821081-AC99-58B3-FD8A-08E0C0F9F22D} = {802C95C5-F8AA-4136-A298-261105286C93}
-		{D6B82683-7DAE-4804-AE1F-477B892E197D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{97E0521A-3EDD-781D-CE28-22CC465D76FA} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{D3765C13-7C11-FEA6-9FD7-3B36F56BAE07} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1EDD6A9D-1770-4385-AADB-26C6AEDD8C12}
