+++ src/Acontplus.Persistence.PostgreSQL/Resources/Images/icon.png
