+++ src/Acontplus.Persistence.PostgreSQL/Utilities/DiagnosticConfig.cs
@@ -1,0 +1,8 @@
+using System.Diagnostics;
+
+namespace Acontplus.Persistence.PostgreSQL.Utilities;
+
+public static class DiagnosticConfig
+{
+    public static readonly ActivitySource ActivitySource = new("Repository");
+}
