+++ src/Acontplus.Persistence.PostgreSQL/Utilities/StringExtensions.cs
@@ -1,0 +1,31 @@
+using System.Text;
+
+namespace Acontplus.Persistence.PostgreSQL.Utilities;
+
+// Extension method for PostgreSQL snake_case naming convention
+public static class StringExtensions
+{
+    public static string ToSnakeCase(this string input)
+    {
+        if (string.IsNullOrEmpty(input))
+            return input;
+
+        var result = new StringBuilder();
+        result.Append(char.ToLowerInvariant(input[0]));
+
+        for (int i = 1; i < input.Length; i++)
+        {
+            if (char.IsUpper(input[i]))
+            {
+                result.Append('_');
+                result.Append(char.ToLowerInvariant(input[i]));
+            }
+            else
+            {
+                result.Append(input[i]);
+            }
+        }
+
+        return result.ToString();
+    }
+}
