--- src/Acontplus.Persistence.Postgres/GlobalUsings.cs
@@ -1,21 +1,0 @@
-global using Acontplus.Core.Abstractions.Persistence;
-global using Acontplus.Core.Domain.Common.Entities;
-global using Acontplus.Core.Domain.Common.Events;
-global using Acontplus.Core.Domain.Enums;
-global using Acontplus.Core.Domain.Exceptions;
-global using Acontplus.Core.DTOs.Requests;
-global using Acontplus.Persistence.Common.Exceptions;
-global using Acontplus.Persistence.Postgres.Ado.Parameters;
-global using Acontplus.Persistence.Postgres.Exceptions;
-global using Microsoft.EntityFrameworkCore;
-global using Microsoft.EntityFrameworkCore.Metadata.Builders;
-global using Microsoft.Extensions.Configuration;
-global using Microsoft.Extensions.Logging;
-global using Npgsql;
-global using NpgsqlTypes;
-global using Polly;
-global using Polly.Retry;
-global using System.Collections.Concurrent;
-global using System.ComponentModel.DataAnnotations.Schema;
-global using System.Data;
-global using System.Data.Common;
