--- src/Acontplus.Persistence.Postgres/Utilities/DiagnosticConfig.cs
@@ -1,8 +1,0 @@
-﻿using System.Diagnostics;
-
-namespace Acontplus.Persistence.Postgres.Utilities;
-
-public static class DiagnosticConfig
-{
-    public static readonly ActivitySource ActivitySource = new("Repository");
-}
